[TEST] Refactor data lake test suite hierarchical
