+++ kyuubi-main/src/test/scala/org/apache/kyuubi/operation/datalake/IcebergOperationSuite.scala
@@ -25,7 +25,7 @@ import org.apache.kyuubi.tags.DataLakeTest
 class IcebergOperationSuite extends WithKyuubiServer with BasicIcebergJDBCTests {
   override protected val conf: KyuubiConf = {
     val kyuubiConf = KyuubiConf().set(KyuubiConf.ENGINE_IDLE_TIMEOUT, 20000L)
-    icebergConfigs.foreach {case (k, v) => kyuubiConf.set(k, v) }
+    icebergConfigs.foreach { case (k, v) => kyuubiConf.set(k, v) }
     kyuubiConf
   }
